requirements.txt
requests >= 2.31.0
pydantic-settings >= 2.1.0
pydantic >= 2.5.3
python-dotenv >= 1.0.0
tweepy >= 4.14.0
Pillow >= 10.1